requirements.txt
Flask=="*"
langchain=="*"
openai=="*"
boto3==1.26.68
faiss-cpu = "*"
tiktoken = "*"